Move test to http client test
